refactor: Place Settings components into subFolder
